scripts/pyproject.toml
[project]
name = "lxr-builds-migration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-storage>=3.0.0",
    "requests>=2.32.3",
]
